chore: improve vertical navigation code sample and remove unnecessary useEffect
