+++ packages/core/src/VerticalNavigation/VerticalNavigation.stories.tsx
@@ -1,6 +1,3 @@
-import { useEffect, useMemo, useState } from "react";
-import { css } from "@emotion/css";
-import { useMediaQuery, useTheme } from "@mui/material";
 import { StoryObj } from "@storybook/react";
 import { expect, userEvent, within } from "@storybook/test";
 import {
@@ -13,19 +10,26 @@ import {
   HvVerticalNavigationTree,
   HvVerticalNavigationTreeView,
   HvVerticalNavigationTreeViewItem,
-  NavigationData,
 } from "@hitachivantara/uikit-react-core";
-import {
-  BarChart,
-  Cloud,
-  Deploy,
-  LogOut,
-  Open,
-  User,
-} from "@hitachivantara/uikit-react-icons";
 
+import { Collapsible as CollapsibleStory } from "./stories/Collapsible";
+import CollapsibleRaw from "./stories/Collapsible?raw";
+import { CollapsibleIcons as CollapsibleIconsStory } from "./stories/CollapsibleIcons";
+import CollapsibleIconsRaw from "./stories/CollapsibleIcons?raw";
+import { CollapsibleIconsWithCustomPopupStyles as CollapsibleIconsWithCustomPopupStylesStory } from "./stories/CollapsibleIconsWithCustomPopupStyles";
+import CollapsibleIconsWithCustomPopupStylesRaw from "./stories/CollapsibleIconsWithCustomPopupStyles?raw";
+import { CollapsibleIconsWithoutSubItems as CollapsibleIconsWithoutSubItemsStory } from "./stories/CollapsibleIconsWithoutSubItems";
+import CollapsibleIconsWithoutSubItemsRaw from "./stories/CollapsibleIconsWithoutSubItems?raw";
+import { Main as MainStory } from "./stories/Main";
+import MainRaw from "./stories/Main?raw";
+import { MobileNavigation as MobileNavigationStory } from "./stories/MobileNavigation";
+import MobileNavigationRaw from "./stories/MobileNavigation?raw";
 import { SliderMode as SliderModeStory } from "./stories/SliderMode";
 import SliderModeRaw from "./stories/SliderMode?raw";
+import { TreeViewMode as TreeViewModeStory } from "./stories/TreeViewMode";
+import TreeViewModeRaw from "./stories/TreeViewMode?raw";
+import { WithoutActions as WithoutActionsStory } from "./stories/WithoutActions";
+import WithoutActionsRaw from "./stories/WithoutActions?raw";
 
 export default {
   title: "Widgets/Vertical Navigation",
@@ -51,85 +55,9 @@ export const Main: StoryObj<HvVerticalNavigationProps> = {
     // Enables Chromatic snapshot
     chromatic: { disableSnapshot: false },
     eyes: { include: true },
+    docs: { source: { code: MainRaw } },
   },
-  render: (args) => {
-    const navigationData = useMemo<NavigationData[]>(
-      () => [
-        { id: "00", label: "Overview" },
-        { id: "01", label: "Analytics" },
-        {
-          id: "02",
-          label: "Storage",
-          data: [
-            {
-              id: "02-01",
-              label: "Cloud",
-              data: [
-                {
-                  id: "02-01-01",
-                  label: "Servers",
-                  href: "https://www.hitachivantara.com/en-us/news.html",
-                },
-                {
-                  id: "02-01-02",
-                  label: "HCP Anywhere",
-                },
-                {
-                  id: "02-01-03",
-                  label: "This Computer",
-                  disabled: true,
-                },
-              ],
-            },
-          ],
-        },
-        {
-          id: "03",
-          label: "Administration",
-          href: "#admin",
-          data: [
-            {
-              id: "03-01",
-              label: "Rest API",
-              href: "#admin-rest",
-              data: [
-                {
-                  id: "03-01-01",
-                  label: "Log Bundle",
-                  href: "#admin-rest-logs",
-                },
-              ],
-            },
-          ],
-        },
-      ],
-      [],
-    );
-
-    const [value, setValue] = useState("00");
-    return (
-      <div style={{ display: "flex", width: 220, height: 530 }}>
-        <HvVerticalNavigation {...args}>
-          <HvVerticalNavigationTree
-            aria-label="Example 1 navigation"
-            selected={value}
-            onChange={(event, data) => {
-              if (data.id === "02-01-01") {
-                event.preventDefault();
-                event.stopPropagation();
-              }
-              setValue(data.id);
-            }}
-            data={navigationData}
-          />
-          <HvVerticalNavigationActions>
-            <HvVerticalNavigationAction label="Profile" icon={<User />} />
-            <HvVerticalNavigationAction label="Logout" icon={<LogOut />} />
-          </HvVerticalNavigationActions>
-        </HvVerticalNavigation>
-      </div>
-    );
-  },
+  render: (args) => <MainStory {...args} />,
 };
 
 export const TreeViewMode: StoryObj<HvVerticalNavigationProps> = {
@@ -140,83 +68,13 @@ export const TreeViewMode: StoryObj<HvVerticalNavigationProps> = {
           "Usage of the [Treeview Design Pattern](https://w3c.github.io/aria-practices/#TreeView) to build a navigation tree for a set of hierarchically organized web pages. " +
           "Instead of TAB, use the arrow keys to navigate through items. Enter performs its default action (i.e. open/close parent nodes, select otherwise).",
       },
+      source: { code: TreeViewModeRaw },
     },
     // Enables Chromatic snapshot
     chromatic: { disableSnapshot: false },
     eyes: { include: true },
   },
-  render: () => {
-    const navigationData = useMemo(
-      () => [
-        { id: "00", label: "Instalation Overview" },
-        {
-          id: "01",
-          label: "Hardware",
-          data: [
-            {
-              id: "01-01",
-              label: "Ambient Monitoring",
-            },
-            {
-              id: "01-02",
-              label: "Server Status Summary",
-            },
-          ],
-        },
-        {
-          id: "02",
-          label: "System",
-          data: [
-            {
-              id: "02-01",
-              label: "Buckets",
-            },
-            {
-              id: "02-02",
-              label: "Admin Users",
-            },
-            {
-              id: "02-03",
-              label: "Log Bundle",
-              data: [
-                {
-                  id: "02-03-01",
-                  label: "Rest API",
-                },
-                {
-                  id: "02-03-02",
-                  label: "License",
-                },
-              ],
-            },
-          ],
-        },
-      ],
-      [],
-    );
-
-    const [value, setValue] = useState("01-01");
-
-    return (
-      <div style={{ display: "flex", width: 220, height: 530 }}>
-        <HvVerticalNavigation>
-          <HvVerticalNavigationTree
-            mode="treeview"
-            collapsible
-            defaultExpanded
-            aria-label="Example 3 navigation"
-            selected={value}
-            onChange={(event, data) => setValue(data.id)}
-            data={navigationData}
-          />
-          <HvVerticalNavigationActions>
-            <HvVerticalNavigationAction label="Profile" icon={<User />} />
-            <HvVerticalNavigationAction label="Logout" icon={<LogOut />} />
-          </HvVerticalNavigationActions>
-        </HvVerticalNavigation>
-      </div>
-    );
-  },
+  render: () => <TreeViewModeStory />,
 };
 
 export const WithoutActions: StoryObj<HvVerticalNavigationProps> = {
@@ -224,92 +82,9 @@ export const WithoutActions: StoryObj<HvVerticalNavigationProps> = {
     // Enables Chromatic snapshot
     chromatic: { disableSnapshot: false },
     eyes: { include: true },
+    docs: { source: { code: WithoutActionsRaw } },
   },
-  render: () => {
-    const navigationData = useMemo(
-      () => [
-        {
-          id: "01",
-          label: "System",
-          data: [
-            {
-              id: "01-01",
-              label: "SCPodF",
-              data: [
-                {
-                  id: "01-01-01",
-                  label: "Compute",
-                },
-                {
-                  id: "01-01-02",
-                  label: "Storage",
-                },
-                {
-                  id: "01-01-03",
-                  label: "Ethernet",
-                },
-                {
-                  id: "01-01-04",
-                  label: "Fiber Channel",
-                  payload: { path: "/hello/world", params: { a: 2, b: "3" } },
-                },
-              ],
-            },
-          ],
-        },
-        {
-          id: "02",
-          label: "Administration",
-          data: [
-            {
-              id: "02-01",
-              label: "Rest API",
-            },
-            {
-              id: "02-02",
-              label: "License",
-            },
-            {
-              id: "02-03",
-              label: "Some big text that shouldn't fit",
-              data: [
-                {
-                  id: "02-03-01",
-                  label: "Rest API",
-                },
-                {
-                  id: "02-03-02",
-                  label: "License",
-                },
-              ],
-            },
-            {
-              id: "02-04",
-              label: "Log Bundle",
-            },
-          ],
-        },
-      ],
-      [],
-    );
-
-    const [value, setValue] = useState("02-03-02");
-
-    return (
-      <div style={{ display: "flex", width: 220 }}>
-        <HvVerticalNavigation>
-          <HvVerticalNavigationTree
-            aria-label="Example 1 navigation"
-            selected={value}
-            onChange={(event, data) => {
-              setValue(data.id);
-            }}
-            data={navigationData}
-          />
-        </HvVerticalNavigation>
-      </div>
-    );
-  },
+  render: () => <WithoutActionsStory />,
 };
 
 export const Collapsible: StoryObj<HvVerticalNavigationProps> = {
@@ -317,96 +92,9 @@ export const Collapsible: StoryObj<HvVerticalNavigationProps> = {
     // Enables Chromatic snapshot
     chromatic: { disableSnapshot: false },
     eyes: { include: true },
+    docs: { source: { code: CollapsibleRaw } },
   },
-  render: () => {
-    const [navigationDataState, setNavigationDataState] = useState<
-      NavigationData[]
-    >([]);
-
-    useEffect(() => {
-      setNavigationDataState([
-        { id: "00", label: "Instalation Overview" },
-        {
-          id: "01",
-          label: "Hardware",
-          icon: <BarChart />,
-          data: [
-            {
-              id: "01-01",
-              label: "Ambient Monitoring",
-            },
-            {
-              id: "01-02",
-              label: "Server Status Summary",
-            },
-          ],
-        },
-        {
-          id: "02",
-          label: "System",
-          data: [
-            {
-              id: "02-01",
-              label: "Buckets",
-            },
-            {
-              id: "02-02",
-              label: "Admin Users",
-            },
-            {
-              id: "02-03",
-              label: "Log Bundle",
-              data: [
-                {
-                  id: "02-03-01",
-                  label: "Rest API",
-                },
-                {
-                  id: "02-03-02",
-                  label: "License",
-                },
-              ],
-            },
-          ],
-        },
-      ]);
-    }, []);
-
-    const [value, setValue] = useState("01-01");
-
-    const [show, setShow] = useState(false);
-
-    const handleIsExpanded = () => {
-      setShow(!show);
-    };
-
-    return (
-      <div style={{ display: "flex", width: 220, height: 530 }}>
-        <HvVerticalNavigation open={show}>
-          <HvVerticalNavigationHeader
-            title="Menu"
-            onCollapseButtonClick={handleIsExpanded}
-            collapseButtonProps={{
-              "aria-label": "collapseButton",
-              "aria-expanded": show,
-            }}
-          />
-          <HvVerticalNavigationTree
-            collapsible
-            defaultExpanded
-            aria-label="Example 3 navigation"
-            selected={value}
-            onChange={(event, data) => setValue(data.id)}
-            data={navigationDataState}
-          />
-          <HvVerticalNavigationActions>
-            <HvVerticalNavigationAction label="Profile" icon={<User />} />
-            <HvVerticalNavigationAction label="Logout" icon={<LogOut />} />
-          </HvVerticalNavigationActions>
-        </HvVerticalNavigation>
-      </div>
-    );
-  },
+  render: () => <CollapsibleStory />,
 };
 
 export const CollapsibleIcons: StoryObj<HvVerticalNavigationProps> = {
@@ -416,132 +104,13 @@ export const CollapsibleIcons: StoryObj<HvVerticalNavigationProps> = {
         story:
           "When collapsed in icon mode only the icons are visible, if an icon is not provided one will be generated based on the first letter of the label.",
       },
+      source: { code: CollapsibleIconsRaw },
     },
     // Enables Chromatic snapshot
     chromatic: { disableSnapshot: false },
     eyes: { include: true },
   },
-  render: () => {
-    const [navigationDataState, setNavigationDataState] = useState<
-      NavigationData[]
-    >([]);
-
-    useEffect(() => {
-      setNavigationDataState([
-        { id: "00", label: "Instalation Overview", icon: <Open /> },
-        {
-          id: "01",
-          label: "Hardware",
-          icon: <BarChart />,
-          data: [
-            {
-              id: "01-01",
-              label: "Ambient Monitoring",
-              icon: <BarChart />,
-            },
-            {
-              id: "01-02",
-              label: "Server Status Summary",
-            },
-          ],
-        },
-        {
-          id: "02",
-          label: "System",
-          icon: <Deploy />,
-          data: [
-            {
-              id: "02-01",
-              label: "Buckets",
-              icon: <Deploy />,
-            },
-            {
-              id: "02-02",
-              label: "Admin Users",
-            },
-            {
-              id: "02-03",
-              label: "Log Bundle",
-              data: [
-                {
-                  id: "02-03-01",
-                  label: "Rest API",
-                },
-                {
-                  id: "02-03-02",
-                  label: "License",
-                },
-              ],
-            },
-          ],
-        },
-        {
-          id: "03",
-          label: "System 2",
-          // icon: <Deploy />,
-          data: [
-            {
-              id: "03-01",
-              label: "Buckets",
-            },
-            {
-              id: "03-02",
-              label: "Admin Users",
-            },
-            {
-              id: "03-03",
-              label: "Log Bundle",
-              data: [
-                {
-                  id: "03-03-01",
-                  label: "Rest API",
-                },
-                {
-                  id: "03-03-02",
-                  label: "License",
-                },
-              ],
-            },
-          ],
-        },
-      ]);
-    }, []);
-
-    const [value, setValue] = useState("01-01");
-
-    const [show, setShow] = useState(false);
-
-    const handleIsExpanded = () => {
-      setShow(!show);
-    };
-
-    return (
-      <div style={{ display: "flex", width: 220, height: 530 }}>
-        <HvVerticalNavigation open={show} useIcons>
-          <HvVerticalNavigationHeader
-            title="Menu"
-            onCollapseButtonClick={handleIsExpanded}
-            collapseButtonProps={{
-              "aria-label": "collapseButton",
-              "aria-expanded": show,
-            }}
-          />
-          <HvVerticalNavigationTree
-            collapsible
-            defaultExpanded
-            aria-label="Example 3 navigation"
-            selected={value}
-            onChange={(event, data) => setValue(data.id)}
-            data={navigationDataState}
-          />
-          <HvVerticalNavigationActions>
-            <HvVerticalNavigationAction label="Profile" icon={<User />} />
-            <HvVerticalNavigationAction label="Logout" icon={<LogOut />} />
-          </HvVerticalNavigationActions>
-        </HvVerticalNavigation>
-      </div>
-    );
-  },
+  render: () => <CollapsibleIconsStory />,
 };
 
 export const CollapsibleIconsWithoutSubItems: StoryObj<HvVerticalNavigationProps> =
@@ -552,72 +121,13 @@ export const CollapsibleIconsWithoutSubItems: StoryObj<HvVerticalNavigationProps
           story:
             "When collapsed in icon mode and no item has sub item, the panel will be have a smaller width.",
         },
+        source: { code: CollapsibleIconsWithoutSubItemsRaw },
       },
       // Enables Chromatic snapshot
       chromatic: { disableSnapshot: false },
       eyes: { include: true },
     },
-    render: () => {
-      const [navigationDataState, setNavigationDataState] = useState<
-        NavigationData[]
-      >([]);
-
-      useEffect(() => {
-        setNavigationDataState([
-          { id: "00", label: "Instalation Overview", icon: <Open /> },
-          {
-            id: "01",
-            label: "Hardware",
-            icon: <BarChart />,
-          },
-          {
-            id: "02",
-            label: "System",
-            icon: <Deploy />,
-          },
-          {
-            id: "03",
-            label: "System 2",
-            icon: <Cloud />,
-          },
-        ]);
-      }, []);
-
-      const [value, setValue] = useState("01-01");
-
-      const [show, setShow] = useState(false);
-
-      const handleIsExpanded = () => {
-        setShow(!show);
-      };
-
-      return (
-        <div style={{ display: "flex", width: 220, height: 530 }}>
-          <HvVerticalNavigation open={show} useIcons>
-            <HvVerticalNavigationHeader
-              title="Menu"
-              onCollapseButtonClick={handleIsExpanded}
-              collapseButtonProps={{
-                "aria-label": "collapseButton",
-                "aria-expanded": show,
-              }}
-            />
-            <HvVerticalNavigationTree
-              collapsible
-              defaultExpanded
-              aria-label="Example 3 navigation"
-              selected={value}
-              onChange={(event, data) => setValue(data.id)}
-              data={navigationDataState}
-            />
-            <HvVerticalNavigationActions>
-              <HvVerticalNavigationAction label="Profile" icon={<User />} />
-              <HvVerticalNavigationAction label="Logout" icon={<LogOut />} />
-            </HvVerticalNavigationActions>
-          </HvVerticalNavigation>
-        </div>
-      );
-    },
+    render: () => <CollapsibleIconsWithoutSubItemsStory />,
   };
 
 export const CollapsibleIconsWithCustomPopupStyles: StoryObj<HvVerticalNavigationProps> =
@@ -628,6 +138,7 @@ export const CollapsibleIconsWithCustomPopupStyles: StoryObj<HvVerticalNavigatio
           story:
             "Custom popup styles can be applied to the popup container by passing a style object to the popupStyles prop.",
         },
+        source: { code: CollapsibleIconsWithCustomPopupStylesRaw },
       },
       // Enables Chromatic snapshot
       chromatic: { disableSnapshot: false },
@@ -638,138 +149,11 @@ export const CollapsibleIconsWithCustomPopupStyles: StoryObj<HvVerticalNavigatio
       const canvas = within(canvasElement);
       const button = canvas.getByRole("button", { name: /collapse/i });
       await userEvent.click(button);
-      await expect(canvas.getAllByRole("listitem")).toHaveLength(4);
+      await expect(
+        canvas.getByRole("heading", { name: /menu/i }),
+      ).toBeInTheDocument();
     },
-    render: () => {
-      const [navigationDataState, setNavigationDataState] = useState<
-        NavigationData[]
-      >([]);
-
-      useEffect(() => {
-        setNavigationDataState([
-          { id: "00", label: "Instalation Overview", icon: <Open /> },
-          {
-            id: "01",
-            label: "Hardware",
-            icon: <BarChart />,
-            data: [
-              {
-                id: "01-01",
-                label: "Ambient Monitoring",
-              },
-              {
-                id: "01-02",
-                label: "Server Status Summary",
-              },
-            ],
-          },
-          {
-            id: "02",
-            label: "System",
-            icon: <Deploy />,
-            data: [
-              {
-                id: "02-01",
-                label: "Buckets",
-                icon: <Deploy />,
-              },
-              {
-                id: "02-02",
-                label: "Admin Users",
-              },
-              {
-                id: "02-03",
-                label: "Log Bundle",
-                data: [
-                  {
-                    id: "02-03-01",
-                    label: "Rest API",
-                  },
-                  {
-                    id: "02-03-02",
-                    label: "License",
-                  },
-                ],
-              },
-            ],
-          },
-          {
-            id: "03",
-            label: "System 2",
-            // icon: <Deploy />,
-            data: [
-              {
-                id: "03-01",
-                label: "Buckets",
-              },
-              {
-                id: "03-02",
-                label: "Admin Users",
-              },
-              {
-                id: "03-03",
-                label: "Log Bundle",
-                data: [
-                  {
-                    id: "03-03-01",
-                    label: "Rest API",
-                  },
-                  {
-                    id: "03-03-02",
-                    label: "License",
-                  },
-                ],
-              },
-            ],
-          },
-        ]);
-      }, []);
-
-      const [value, setValue] = useState("01-01");
-
-      const [show, setShow] = useState(false);
-
-      const handleIsExpanded = () => {
-        setShow(!show);
-      };
-
-      const popupStyles = css({
-        padding: "20px",
-        backgroundColor: "coral",
-
-        "& > div": {
-          border: "3px solid orange",
-        },
-      });
-
-      return (
-        <div style={{ display: "flex", width: 220, height: 530 }}>
-          <HvVerticalNavigation open={show} useIcons>
-            <HvVerticalNavigationHeader
-              title="Menu"
-              onCollapseButtonClick={handleIsExpanded}
-              collapseButtonProps={{
-                "aria-label": "collapseButton",
-                "aria-expanded": show,
-              }}
-            />
-            <HvVerticalNavigationTree
-              collapsible
-              defaultExpanded
-              aria-label="Example 3 navigation"
-              selected={value}
-              onChange={(event, data) => setValue(data.id)}
-              data={navigationDataState}
-              classes={{ navigationPopup: popupStyles }}
-            />
-            <HvVerticalNavigationActions>
-              <HvVerticalNavigationAction label="Profile" icon={<User />} />
-              <HvVerticalNavigationAction label="Logout" icon={<LogOut />} />
-            </HvVerticalNavigationActions>
-          </HvVerticalNavigation>
-        </div>
-      );
-    },
+    render: () => <CollapsibleIconsWithCustomPopupStylesStory />,
   };
 
 export const SliderMode: StoryObj<HvVerticalNavigationProps> = {
@@ -789,127 +173,11 @@ export const MobileNavigation: StoryObj<HvVerticalNavigationProps> = {
         story:
           "Example of an implementation of the Design System Mobile Navigation pattern: the Vertical Navigation component changes to Slider mode when the window is in a smaller size.",
       },
+      source: { code: MobileNavigationRaw },
     },
     // Enables Chromatic snapshot
     chromatic: { disableSnapshot: false },
     eyes: { include: true },
   },
-  render: () => {
-    const [navigationDataState, setNavigationDataState] = useState<
-      NavigationData[]
-    >([]);
-
-    useEffect(() => {
-      setNavigationDataState([
-        {
-          id: "menu1",
-          label: "Menu 1",
-          path: "",
-          icon: <Open />,
-          data: [
-            {
-              id: "menu1-1",
-              label: "Menu 1-1",
-              path: "",
-              icon: <Open />,
-              parent: null,
-            },
-            {
-              id: "menu1-2",
-              label: "Menu 1-2",
-              path: "",
-              icon: <BarChart />,
-              data: [
-                {
-                  id: "menu1-2-1",
-                  label: "Menu 1-2-1",
-                  path: "",
-                  icon: <Open />,
-                  parent: null,
-                },
-                {
-                  id: "menu1-2-2",
-                  label: "Menu 1-2-2",
-                  path: "",
-                  icon: <BarChart />,
-                  parent: null,
-                },
-                {
-                  id: "menu1-2-3",
-                  label: "Menu 1-2-3",
-                  path: "",
-                  icon: <Deploy />,
-                  parent: null,
-                },
-              ],
-              parent: null,
-            },
-            {
-              id: "menu1-3",
-              label: "Menu 1-3",
-              path: "",
-              icon: <Deploy />,
-              parent: null,
-            },
-          ],
-          parent: null,
-        },
-        {
-          id: "menu2",
-          label: "Menu 2",
-          path: "",
-          icon: <BarChart />,
-          parent: null,
-        },
-        {
-          id: "menu3",
-          label: "Menu 3",
-          path: "",
-          icon: <Deploy />,
-          parent: null,
-        },
-      ]);
-    }, []);
-
-    const [value, setValue] = useState("menu1-3");
-
-    const [show, setShow] = useState(true);
-
-    const theme = useTheme();
-
-    const isXs = useMediaQuery(theme.breakpoints.down("sm"));
-
-    const handleIsExpanded = () => {
-      setShow(!show);
-    };
-
-    return (
-      <div>
-        <div style={{ display: "flex", width: 220, height: 530 }}>
-          <HvVerticalNavigation open={show} slider={isXs}>
-            <HvVerticalNavigationHeader
-              title="Menu"
-              onCollapseButtonClick={handleIsExpanded}
-              collapseButtonProps={{
-                "aria-label": "collapseButton",
-                "aria-expanded": show,
-              }}
-            />
-            <HvVerticalNavigationTree
-              collapsible
-              defaultExpanded
-              aria-label="Example 4 navigation slider"
-              selected={value}
-              onChange={(event, data) => setValue(data.id)}
-              data={navigationDataState}
-            />
-            <HvVerticalNavigationActions>
-              <HvVerticalNavigationAction label="Profile" icon={<User />} />
-              <HvVerticalNavigationAction label="Logout" icon={<LogOut />} />
-            </HvVerticalNavigationActions>
-          </HvVerticalNavigation>
-        </div>
-      </div>
-    );
-  },
+  render: () => <MobileNavigationStory />,
 };
